// src/shared/enums/file-types.enum.ts
export enum FileType {
  CHALLENGE_LOGO = 'CHALLENGE_LOGO',
  SCHOOL_LOGO = 'SCHOOL_LOGO',
  BUSINESS_LEARNING_FILE = 'BUSINESS_LEARNING_FILE',
  PITCH_FEEDBACK_FILE = 'PITCH_FEEDBACK_FILE',
  BUSINESS_MODEL_PDF = 'BUSINESS_MODEL_PDF',
  USER_AVATARS = 'USER_AVATARS',
  BULK_UPLOAD_STUDENTS = 'BULK_UPLOAD_STUDENTS',
  COUNTRY_AVATARS = 'COUNTRY_AVATARS',
}